+++ coloraide_extras/spaces/cmyk.py
@@ -3,8 +3,9 @@ Uncalibrated, naive CMYK color space.
 
 https://www.w3.org/TR/css-color-5/#cmyk-rgb
 """
-from coloraide.spaces import Space, RE_DEFAULT_MATCH, GamutBound, WHITES
-import re
+from coloraide.spaces import Space
+from coloraide.gamut.bounds import GamutBound
+from coloraide.cat import WHITES
 from coloraide.util import MutableVector
 from typing import Tuple
 
@@ -47,7 +48,6 @@ class CMYK(Space):
         "yellow": 'y',
         "black": 'k'
     }
-    DEFAULT_MATCH = re.compile(RE_DEFAULT_MATCH.format(color_space='|'.join(SERIALIZE), channels=4))
     WHITE = WHITES['2deg']['D65']
 
     BOUNDS = (