+++ coloraide_extras/spaces/prismatic.py
@@ -6,8 +6,9 @@ Creates a Maxwell color triangle with a lightness component.
 http://psgraphics.blogspot.com/2015/10/prismatic-color-model.html
 https://studylib.net/doc/14656976/the-prismatic-color-space-for-rgb-computations
 """
-from coloraide.spaces import Space, RE_DEFAULT_MATCH, GamutBound, WHITES
-import re
+from coloraide.spaces import Space
+from coloraide.gamut.bounds import GamutBound
+from coloraide.cat import WHITES
 from coloraide.util import MutableVector
 from typing import Tuple
 
@@ -42,7 +43,6 @@ class Prismatic(Space):
         "green": 'g',
         "blue": 'b'
     }
-    DEFAULT_MATCH = re.compile(RE_DEFAULT_MATCH.format(color_space='|'.join(SERIALIZE), channels=4))
     WHITE = WHITES['2deg']['D65']
 
     BOUNDS = (