+++ coloraide_extras/spaces/xyy.py
@@ -3,8 +3,9 @@ The xyY color space.
 
 https://en.wikipedia.org/wiki/CIE_1931_color_space#CIE_xy_chromaticity_diagram_and_the_CIE_xyY_color_space
 """
-from coloraide.spaces import Space, RE_DEFAULT_MATCH, GamutUnbound, WHITES
-import re
+from coloraide.spaces import Space
+from coloraide.gamut.bounds import GamutUnbound
+from coloraide.cat import WHITES
 from coloraide import util
 from coloraide.util import MutableVector
 from typing import Tuple
@@ -17,7 +18,6 @@ class XyY(Space):
     NAME = "xyy"
     SERIALIZE = ("--xyy",)  # type: Tuple[str, ...]
     CHANNEL_NAMES = ("x", "y", "Y")
-    DEFAULT_MATCH = re.compile(RE_DEFAULT_MATCH.format(color_space='|'.join(SERIALIZE), channels=3))
     WHITE = WHITES['2deg']['D65']
 
     BOUNDS = (